improvement(Passkey): Accept decimal values of cred params algorithm and round it to an int

+++ common/src/commonMain/kotlin/com/artemchep/keyguard/common/service/passkey/entity/CreatePasskeyRequest.kt
@@ -73,7 +73,7 @@ data class CreatePasskeyUser(
 @Serializable
 data class CreatePasskeyPubKeyCredParams(
     // https://www.iana.org/assignments/cose/cose.xhtml#algorithms
-    val alg: Int,
+    val alg: Double,
     val type: String,
 )
